Extract class/OOP builtins from vm.c into vm_builtin_class.c
Move 17 class introspection builtins (get_class, class_exists, is_a,
call_user_func, etc.) and their helpers into a dedicated file.

Three helpers shared with the bytecode interpreter are promoted from
static to PH7_PRIVATE: VmInstanceOf, VmClassMemberAccess, and
VmExtractClassFromValue. VmFrame struct and VmLocalExec/VmErrorFormat
are also moved to ph7int.h to support cross-file access.

Builds and tests pass on all platforms (GNU full/tiny, MSVC).
